requirements-gpu.txt
diffusers[torch]
opencv-python>=4.0
transformers==4.45.2
accelerate==1.0.1
scipy==1.13.1
safetensors==0.4.5
matplotlib==3.8.4
jupyterlab>=4.2.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.2
lightning==2.2.2
torchmetrics
tensorboard==2.16.2
pillow==11.0.0
torch
torchvision
seaborn==0.13.1
scikit-image==0.24.0
wandb==0.18.6
hydra-core==1.3.2